Implemented most of the search.
A difficult non-implemented case is when one backing directory has
substring search for the id, and uses prefix/suffix. Then we cannot pass
a massaged search string, but we would have to search on all ids and
then post-filter the search.

So it's not a good idea to do substring search on the id when prefixes
are in use.