Made changes to be able to use Local Directory with an Indirect directory that would be created automatically.
Fixed a bug related to the fact that the admin is not always a member
of the site, but should be able to access to the directory properties.

WARNING: directory_id does not refer to the id of the directory created
locally (this one will have the same id than the directory with type
'CPS Local Directory'), but is used when creating an indirect directory
locally : it is the id of the directory it refers to (entries will be links
to entries of that directory).